DPMA
# DPMA
A model-averaging approach for dynamic prediction of an event using multiple longitudinal markers
Dynamic event  prediction, using joint modeling of survival time and longitudinal variables, is extremely useful in personalized medicine. However, the estimation of joint models including many longitudinal markers is still a computational challenge because of the high number of random effects and parameters to be estimated. In this paper, we propose a model averaging strategy to combine predictions from several joint models for the event, including one longitudinal marker only   or pairwise longitudinal markers. The prediction is computed as the weighted mean of the predictions from the one-marker or two-marker  models, with the time-dependent weights estimated by minimizing the time-dependent Brier score. This method enables us to combine a large number of predictions issued from joint models to achieve a reliable and accurate individual prediction. Advantages and limits of the proposed methods are highlighted in a simulation study by comparison with the predictions from well-specified and misspecified all-marker joint models as well as the one-marker  and two-marker  joint models. Using the PBC2 data set, the method is used to predict the risk of death in patients with primary biliary cirrhosis.

## References:
Reza Hashemi, Taban Baghfalaki, Viviane Philipps and Helene Jacqmin-Gadda (2024). Dynamic prediction of an event using multiple longitudinal markers: a model averaging approach, Statistics in Medicine.